Add version prefixes option to released-version goal
Allow running the released-version goal with optional parameter
versionPrefix, which filters the matched released versions to versions
that begin with that prefix.

For example, if the released versions are 1.3.1, 1.3.2, 1.4.1, 1.4.2,
and you run `mvn build-helper:released-version` normally, it will
find version 1.4.2 as the latest released version.

This commit makes it so if you give it the parameter versionPrefix=1.3,
then it will find 1.3.2 instead.

You can also pass in the versionPrefix via a property:

   mvn build-helper:released-version -DreleasedVersion.versionPrefix=1.3