Add skipIfMissing flag
- build-helper:add-source
- build-helper:add-test-source
- build-helper:add-resource
- build-helper:add-test-resource

add flag `<skipIfMissing>true|false<skipIfMissing>`. If this flag is set
to true and a directory that should be added does not exist, skip it and
do not add it to the build. This allows a multi module build to use
these plugins unconditionally. E.g.

```
<plugin>
    <groupId>org.codehaus.mojo</groupId>
    <artifactId>build-helper-maven-plugin</artifactId>
    <executions>
        <execution>
            <id>kotlin</id>
            <goals>
                <goal>add-source</goal>
            </goals>
            <phase>generate-sources</phase>
            <configuration>
                <skipIfMissing>true</skipIfMissing>
                <sources>
                    <source>${project.basedir}/src/main/kotlin</source>
                </sources>
            </configuration>
        </execution>
    </executions>
</plugin>
```

can be configured unconditionally and only modules that actually have a
`src/main/kotlin` folder will add this as a compile root.